Add QR Code and mobile icons
