+++ packages/@coorpacademy-nova-icons/icons.js
@@ -56,11 +56,12 @@ const icons: Array<Icon> = [
   {filePath: path.resolve('./third-party/nova-line.iconjar/icons/interface-alert-circle.svg')},
   {filePath: path.resolve('./third-party/nova-line.iconjar/icons/key-1.svg')},
   {filePath: path.resolve('./third-party/nova-line.iconjar/icons/pencil-2.svg')},
-  {filePath: path.resolve('./third-party/nova-line.iconjar/icons/qr-code.svg')},
+  {filePath: path.resolve('./third-party/nova-line.iconjar/icons/qr-code.1.svg')},
   {filePath: path.resolve('./third-party/nova-line.iconjar/icons/remove-circle-1.svg')},
   {filePath: path.resolve('./third-party/nova-line.iconjar/icons/video-clip-3.svg')},
   {filePath: path.resolve('./third-party/nova-line.iconjar/icons/link-broken.svg')},
   {filePath: path.resolve('./third-party/nova-line.iconjar/icons/ring-planet.svg')},
+  {filePath: path.resolve('./third-party/nova-line.iconjar/icons/network-mobile.svg')},
   {filePath: path.resolve('./third-party/nova-solid.iconjar/icons/add-1.svg')},
   {filePath: path.resolve('./third-party/nova-solid.iconjar/icons/audio-control-play.svg')},
   {filePath: path.resolve('./third-party/nova-solid.iconjar/icons/bookmark-3.svg')},
