+++ packages/@coorpacademy-nova-icons/src/index.js
@@ -41,17 +41,18 @@ import _NovaCompositionCoorpacademyListBullets3 from './components/nova-composit
 import _NovaCompositionCoorpacademyLightbulb from './components/nova-composition/coorpacademy/lightbulb';
 import _NovaCompositionCoorpacademyClock from './components/nova-composition/coorpacademy/clock';
 import _NovaLineBusinessCircleView from './components/nova-line/business/circle-view';
-import _NovaLineShoppingQrCode from './components/nova-line/shopping/qr-code';
 import _NovaLineLoginKey1 from './components/nova-line/login/key-1';
 import _NovaLineContentEditionLinkBroken from './components/nova-line/content-edition/link-broken';
 import _NovaLineContentEditionPencil2 from './components/nova-line/content-edition/pencil-2';
 import _NovaLineInterfaceFeedbackInterfaceAlertCircle from './components/nova-line/interface-feedback/interface-alert-circle';
 import _NovaLineStatusCheckCircle1 from './components/nova-line/status/check-circle-1';
 import _NovaLineStatusCloseCircle from './components/nova-line/status/close-circle';
+import _NovaLineNetworkNetworkMobile from './components/nova-line/network/network-mobile';
 import _NovaLineRemoveAddRemoveCircle1 from './components/nova-line/remove-add/remove-circle-1';
 import _NovaLineBookmarksTagsBookmark5 from './components/nova-line/bookmarks-tags/bookmark-5';
 import _NovaLineAudioAudioControlPlay from './components/nova-line/audio/audio-control-play';
 import _NovaLineVideosVideoClip3 from './components/nova-line/videos/video-clip-3';
+import _NovaLineMobilephoneQrCode1 from './components/nova-line/mobilephone/qr-code.1';
 import _NovaLineFilesOfficeFileOfficePdf from './components/nova-line/files-office/file-office-pdf';
 import _NovaLineComputersComputerScreen1 from './components/nova-line/computers/computer-screen-1';
 import _NovaLineSpaceRingPlanet from './components/nova-line/space/ring-planet';
@@ -120,17 +121,18 @@ export const NovaCompositionCoorpacademyListBullets3: Icon = _NovaCompositionCoo
 export const NovaCompositionCoorpacademyLightbulb: Icon = _NovaCompositionCoorpacademyLightbulb;
 export const NovaCompositionCoorpacademyClock: Icon = _NovaCompositionCoorpacademyClock;
 export const NovaLineBusinessCircleView: Icon = _NovaLineBusinessCircleView;
-export const NovaLineShoppingQrCode: Icon = _NovaLineShoppingQrCode;
 export const NovaLineLoginKey1: Icon = _NovaLineLoginKey1;
 export const NovaLineContentEditionLinkBroken: Icon = _NovaLineContentEditionLinkBroken;
 export const NovaLineContentEditionPencil2: Icon = _NovaLineContentEditionPencil2;
 export const NovaLineInterfaceFeedbackInterfaceAlertCircle: Icon = _NovaLineInterfaceFeedbackInterfaceAlertCircle;
 export const NovaLineStatusCheckCircle1: Icon = _NovaLineStatusCheckCircle1;
 export const NovaLineStatusCloseCircle: Icon = _NovaLineStatusCloseCircle;
+export const NovaLineNetworkNetworkMobile: Icon = _NovaLineNetworkNetworkMobile;
 export const NovaLineRemoveAddRemoveCircle1: Icon = _NovaLineRemoveAddRemoveCircle1;
 export const NovaLineBookmarksTagsBookmark5: Icon = _NovaLineBookmarksTagsBookmark5;
 export const NovaLineAudioAudioControlPlay: Icon = _NovaLineAudioAudioControlPlay;
 export const NovaLineVideosVideoClip3: Icon = _NovaLineVideosVideoClip3;
+export const NovaLineMobilephoneQrCode1: Icon = _NovaLineMobilephoneQrCode1;
 export const NovaLineFilesOfficeFileOfficePdf: Icon = _NovaLineFilesOfficeFileOfficePdf;
 export const NovaLineComputersComputerScreen1: Icon = _NovaLineComputersComputerScreen1;
 export const NovaLineSpaceRingPlanet: Icon = _NovaLineSpaceRingPlanet;
